This PR adds a ChipV1Clusters.json directly under zcl-builtin/dotdot.
This is not the best place for this file to live since it is a specific
configuration. But it is helpful for exploration purposes. Basically this
file is the file generated by the UI when running 'npm run zapc' and choosing
the OnOff device.
Then you could save this file using the windows menu.