Adds a chip-tool-command.zapt template to generate the chip-tool commands
This PR add a new template file, as well as an helper method for the templating
system that transform a string into a different version more suitable for
command line.

At the moment the chip-tool template is does not fully work. For some reasons
it still misses some commands for some clusters. It may be related to the
ChipV1Clusters.json configuration as those commands may not be activated.
The place where those commands are declared to appears in the UI is inside
ChipDevice.xml, which seems fine.

Also the template does not handle the body of the EncodeCommand method, since I
need to figure out how to access the fields of the methods. I would expect
the necessary bits to live in src-electron/generator/helper-session.js

Similarly the current template does not take into account the cluster attributes
in order to generate the right read command.
I would expect user_cluster_attributes from src-electron/generator/helper-session.js
to be the right templating method to call for that.